[Silver III] Title: 주유소, Time: 384 ms, Memory: 36132 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -8,21 +8,21 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
         int n = Integer.parseInt(br.readLine());
-        int city[] = new int[n - 1];
-        int cost[] = new int[n];
+        long city[] = new long[n - 1];
+        long cost[] = new long[n];
 
         StringTokenizer st = new StringTokenizer(br.readLine(), " ");
         for (int i = 0; i < city.length; i++) {
-            city[i] = Integer.parseInt(st.nextToken());
+            city[i] = Long.parseLong(st.nextToken());
         }
 
         st = new StringTokenizer(br.readLine(), " ");
         for (int i = 0; i < cost.length; i++) {
-            cost[i] = Integer.parseInt(st.nextToken());
+            cost[i] = Long.parseLong(st.nextToken());
         }
-        
-        int sum = cost[0] * city[0];    //  초기값 (처음엔 주유해야함)
-        int mincost = cost[0];  //  두번째부터 최소 비용 정리
+
+        long sum = cost[0] * city[0];    //  초기값 (처음엔 주유해야함)
+        long mincost = cost[0];  //  두번째부터 최소 비용 정리
 
         for (int i = 1; i < city.length; i++) {
             if (mincost > cost[i]) {
